api: Promote idhash supplemental API to core
