+++ src/core/idhash_test.c
@@ -1,5 +1,5 @@
 //
-// Copyright 2024 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2025 Staysail Systems, Inc. <info@staysail.tech>
 //
 // This software is supplied under the terms of the MIT License, a
 // copy of which should be located in the distribution where this
@@ -8,8 +8,6 @@
 //
 
 #include <nuts.h>
-
-#include <nng/supplemental/util/idhash.h>
 
 void
 test_id_basic(void)
