+++ src/supplemental/util/CMakeLists.txt
@@ -7,9 +7,7 @@
 # found online at https://opensource.org/licenses/MIT.
 #
 
-nng_sources(idhash.c options.c)
+nng_sources(options.c)
 nng_headers(
-	nng/supplemental/util/idhash.h
 	nng/supplemental/util/options.h)
-nng_test(idhash_test)
 nng_test(options_test)
